requirements.txt
PyYAML
requests>=2
twine
black
build
isort
